+++ BrazeProject/android/app/src/main/java/com/brazeproject/MainApplication.kt
@@ -1,0 +1,52 @@
+package com.brazeproject
+
+import android.app.Application
+import android.util.Log
+import com.braze.BrazeActivityLifecycleCallbackListener
+import com.braze.support.BrazeLogger.logLevel
+import com.facebook.react.PackageList
+import com.facebook.react.ReactApplication
+import com.facebook.react.ReactNativeHost
+import com.facebook.react.ReactPackage
+import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.load
+import com.facebook.react.defaults.DefaultReactNativeHost
+import com.facebook.soloader.SoLoader
+
+class MainApplication : Application(), ReactApplication {
+    private val reactNativeHost = object: DefaultReactNativeHost(this) {
+        override fun getUseDeveloperSupport(): Boolean {
+            return BuildConfig.DEBUG
+        }
+
+        override fun getPackages(): List<ReactPackage> {
+            // Packages that cannot be autolinked yet can be added manually here, for example:
+            // packages.add(new MyReactNativePackage());
+            return PackageList(this).packages
+        }
+
+        override fun getJSMainModuleName(): String {
+            return "index"
+        }
+
+        override val isNewArchEnabled: Boolean
+            get() = BuildConfig.IS_NEW_ARCHITECTURE_ENABLED
+        override val isHermesEnabled: Boolean
+            get() = BuildConfig.IS_HERMES_ENABLED
+    }
+
+    override fun getReactNativeHost(): ReactNativeHost {
+        return reactNativeHost
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        registerActivityLifecycleCallbacks(BrazeActivityLifecycleCallbackListener())
+        logLevel = Log.VERBOSE
+        SoLoader.init(this,false)
+        if (BuildConfig.IS_NEW_ARCHITECTURE_ENABLED) {
+            // If you opted-in for the New Architecture, we load the native entry point for this app.
+            load()
+        }
+        ReactNativeFlipper.initializeFlipper(this, reactNativeHost.reactInstanceManager)
+    }
+}